+++ atuin-client/src/settings/behaviour.rs
@@ -1,0 +1,86 @@
+use clap::ValueEnum;
+use serde::Deserialize;
+
+use super::Settings;
+
+// Exit
+
+#[derive(Clone, Debug, Deserialize, Copy)]
+pub enum ExitMode {
+    #[serde(rename = "return-original")]
+    ReturnOriginal,
+
+    #[serde(rename = "return-query")]
+    ReturnQuery,
+}
+
+// Filter
+
+#[derive(Clone, Debug, Deserialize, Copy, PartialEq, Eq, ValueEnum)]
+pub enum FilterMode {
+    #[serde(rename = "global")]
+    Global = 0,
+
+    #[serde(rename = "host")]
+    Host = 1,
+
+    #[serde(rename = "session")]
+    Session = 2,
+
+    #[serde(rename = "directory")]
+    Directory = 3,
+
+    #[serde(rename = "workspace")]
+    Workspace = 4,
+}
+
+impl FilterMode {
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            FilterMode::Global => "GLOBAL",
+            FilterMode::Host => "HOST",
+            FilterMode::Session => "SESSION",
+            FilterMode::Directory => "DIRECTORY",
+            FilterMode::Workspace => "WORKSPACE",
+        }
+    }
+}
+
+// Search
+
+#[derive(Clone, Debug, Deserialize, Copy, ValueEnum, PartialEq)]
+pub enum SearchMode {
+    #[serde(rename = "prefix")]
+    Prefix,
+
+    #[serde(rename = "fulltext")]
+    #[clap(aliases = &["fulltext"])]
+    FullText,
+
+    #[serde(rename = "fuzzy")]
+    Fuzzy,
+
+    #[serde(rename = "skim")]
+    Skim,
+}
+
+impl SearchMode {
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            SearchMode::Prefix => "PREFIX",
+            SearchMode::FullText => "FULLTXT",
+            SearchMode::Fuzzy => "FUZZY",
+            SearchMode::Skim => "SKIM",
+        }
+    }
+    pub fn next(&self, settings: &Settings) -> Self {
+        match self {
+            SearchMode::Prefix => SearchMode::FullText,
+            // if the user is using skim, we go to skim
+            SearchMode::FullText if settings.search_mode == SearchMode::Skim => SearchMode::Skim,
+            // otherwise fuzzy.
+            SearchMode::FullText => SearchMode::Fuzzy,
+            SearchMode::Fuzzy | SearchMode::Skim => SearchMode::Prefix,
+        }
+    }
+}
