+++ atuin-client/src/settings/sync.rs
@@ -1,0 +1,6 @@
+use serde::Deserialize;
+
+#[derive(Clone, Debug, Deserialize, Default)]
+pub struct Sync {
+    pub records: bool,
+}
